requirements.txt
pyautogen
pyautogen[retrievechat]
python-dotenv